Refactor B0shim tab
